feat: add laporan

+++ src/lib/prioitem.py
@@ -3,6 +3,6 @@ from typing import Any
 
 @dataclass(order=True)
 class PrioritizedItem:
-    cost: int
+    cost: float
     node: Any=field(compare=False)
     path: Any=field(compare=False)
